purge: fix wipe issues
When a cluster is purged, there is leftover on partitions since they are
not wiped properly and this prevent from redeploying a cluster, indeed, the osd
disk activate process will complain about mismatching uuids.

In order to fix this isse, we must wipe the first 4k of each created
partitions of the OSD device.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>